fix warnings

+++ examples/demos/check_box_appearance/src/check_box_appearance.cpp
@@ -131,7 +131,7 @@ public:
     label_width.location({10, 293});
     
     numeric_up_down_width.value_changed += [&] {
-      check_box_demo.width(numeric_up_down_width.value());
+      check_box_demo.width(static_cast<int>(numeric_up_down_width.value()));
     };
     numeric_up_down_width.value(check_box_demo.width());
     numeric_up_down_width.set_range(0, std::numeric_limits<int32_t>::max());
@@ -144,7 +144,7 @@ public:
     label_height.location({10, 328});
     
     numeric_up_down_height.value_changed += [&] {
-      check_box_demo.height(numeric_up_down_height.value());
+      check_box_demo.height(static_cast<int>(numeric_up_down_height.value()));
     };
     numeric_up_down_height.value(check_box_demo.height());
     numeric_up_down_height.set_range(0, std::numeric_limits<int32_t>::max());
